+++ NEWS.md
@@ -1,7 +1,15 @@
 # pxmake (development version)
 
+# pxmake 0.9.0
+
+## New features
+- Let `metamake` accept a data frame as input and create a minimal metadata 
+template for the data.
+
 ## Bug fixes and minor improvements
 - Handle px-files with either zero heading variables or zero stub variables. (#95)
+- Fix edge case in list sorting caused by px-files with no codes.
+- Use values as codes if no code is given for add_totals variable.
 
 # pxmake 0.8.1